requirements.txt
Flask==3.1.3
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
psycopg2-binary==2.9.10
python-dotenv>=1.0.0
gunicorn>=23.0.0
flasgger==0.9.7.1
Werkzeug==3.1.3
SQLAlchemy==2.0.40
itsdangerous==2.2.0
click==8.1.8
